third_party/WebKit/LayoutTests/http/tests/security/xssAuditor/form-action-expected.txt

CONSOLE ERROR: line 4: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?q=%3Cform%20action=http://127.0.0.1:8000/foo&clutter=bar%20method=x%3E%3Cinput%20type=submit%3E%3Cinput%20name=x%20value=%27Please%20type%20your%20PIN.%27%3E&notifyDone=1&showAction=1' because its source code was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
ALERT: Form action set to data:,
